+++ src/test/java/racing/domain/RacingCarTest.java
@@ -1,11 +1,11 @@
-package racing;
+package racing.domain;
 
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racing.domain.RacingCar;
+import racing.FixedRandomForwardStrategy;
 import racing.strategy.MoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
