+++ src/test/java/racing/view/InputViewTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.view;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
